Implemented Richard's feedback
