+++ Modules/OrgUnit/classes/Extension/class.ilOrgUnitExtension.php
@@ -14,6 +14,7 @@ abstract class ilOrgUnitExtension extends ilObjectPlugin {
 	 * @var ilObjOrgUnitTree
 	 */
 	protected $ilObjOrgUnitTree;
+
 	/**
 	 * @var int
 	 */
@@ -31,6 +32,28 @@ abstract class ilOrgUnitExtension extends ilObjectPlugin {
 		parent::__construct($a_ref_id);
 		$this->ilObjOrgUnitTree = ilObjOrgUnitTree::_getInstance();
 		$this->parent_ref_id = $tree->getParentId($a_ref_id ? $a_ref_id : $_GET['ref_id']);
+	}
+
+	/**
+	 * Returns all Orgu Plugin Ids of active plugins where the Plugin wants to be shown in the tree. ($plugin->showInTree() == true)
+	 *
+	 * @return string[]
+	 */
+	public static function getActivePluginIdsForTree() {
+		/**
+		 * @var $plugin ilOrgUnitExtensionPlugin
+		 */
+		$list = array();
+
+		$plugin_ids = ilPlugin::getActivePluginIdsForSlot(IL_COMP_MODULE, "OrgUnit", "orguext");
+		foreach ($plugin_ids as $plugin_id) {
+			$plugin = ilPlugin::getRepoPluginObjectByType($plugin_id);
+			if ($plugin->showInTree()) {
+				$list[] = $plugin_id;
+			}
+		}
+
+		return $list;
 	}
 
 
@@ -51,42 +74,24 @@ abstract class ilOrgUnitExtension extends ilObjectPlugin {
 
 
 	/**
-	 * @param int $ref_id       returns all employees of the given org unit.
+	 * Get all user ids of employees of the underlying OrgUnit.
+	 *
 	 * @param bool $recursively include all employees in the suborgunits
 	 * @return int[]
 	 */
-	public function getEmployees($ref_id, $recursively = false) {
-		return $this->ilObjOrgUnitTree->getEmployees($ref_id, $recursively);
+	public function getEmployees($recursively = false) {
+		return $this->ilObjOrgUnitTree->getEmployees($this->parent_ref_id, $recursively);
 	}
 
 
 	/**
-	 * Get the IDs of the employees of the org unit this plugin belongs to.
+	 * Get all user ids of superiors of the underlying OrgUnit
 	 *
 	 * @param bool $recursively
 	 * @return int[]
 	 */
-	public function getMyEmployees($recursively = false) {
-		return $this->getEmployees($this->parent_ref_id, $recursively);
-	}
-
-
-	/**
-	 * @param $ref_id
-	 * @param bool $recursively
-	 * @return int[]
-	 */
-	public function getSuperiors($ref_id, $recursively = false) {
-		return $this->ilObjOrgUnitTree->getSuperiors($ref_id, $recursively);
-	}
-
-
-	/**
-	 * @param bool $recursively
-	 * @return int[]
-	 */
-	public function getMySuperiors($recursively = false) {
-		return $this->getSuperiors($this->parent_ref_id, $recursively);
+	public function getSuperiors($recursively = false) {
+		return $this->ilObjOrgUnitTree->getSuperiors($this->parent_ref_id, $recursively);
 	}
 
 
@@ -94,6 +99,6 @@ abstract class ilOrgUnitExtension extends ilObjectPlugin {
 	 * @return ilObjOrgUnit
 	 */
 	public function getOrgUnit() {
-		return new ilObjOrgUnit($this->parent_ref_id);
+		return ilObjectFactory::getInstanceByRefId($this->parent_ref_id);
 	}
 }
